requirements.txt
absl-py
alignment
altgraph
appdirs
async-timeout
audioread
backcall
blinker
brotlipy
cachetools
certifi
cffi
chardet
click
cycler
Cython
cytoolz
decorator
Distance
einops~=0.8.0
et-xmlfile
fsspec
future
g2p-en
g2pM~=0.1.2.5
google-auth
google-auth-oauthlib
grpcio
h5py~=3.11.0
horology
httplib2
idna
imageio
inflect
ipdb
ipython
ipython-genutils
jdcal
jedi
jieba~=0.42.1
jiwer
joblib
kiwisolver
librosa~=0.9.2
llvmlite
Markdown
matplotlib~=3.7.5
miditoolkit
mido
music21
networkx
nltk~=3.8.1
numba
numpy~=1.24.4
oauth2client
oauthlib
olefile
packaging
pandas~=2.0.3
parso
patsy
pexpect
pickleshare
Pillow
pooch
praat-parselmouth
prompt-toolkit
protobuf
ptyprocess
pyasn1
pyasn1-modules
pycparser
pycwt~=0.4.0b0
Pygments
PyInstaller
PyJWT
pyloudnorm~=0.1.1
pyparsing
pypinyin~=0.51.0
PySocks
python-dateutil
python-Levenshtein
pytorch-lightning
pytz
PyWavelets
pyworld
PyYAML~=6.0.1
regex
requests
requests-oauthlib
resampy
Resemblyzer~=0.1.4
rsa
scikit-image~=0.21.0
scikit-learn~=1.3.2
scipy~=1.10.1
six~=1.16.0
SoundFile
stopit
tensorboard
tensorboard-plugin-wit
tensorboardX
TextGrid~=1.6.1
threadpoolctl
toolz
torch~=2.3.0
torchaudio
torchvision
tqdm~=4.66.4
traitlets
typing
urllib3
uuid
wcwidth
webencodings
webrtcvad~=2.0.10
Werkzeug
pretty-midi
gradio~=4.29.0
lightning~=2.2.4